% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{mx.api-package}
\alias{mx.api}
\alias{mx.api-package}
\title{\packageTitle{mx.api}}
\description{\packageDescription{mx.api}}
\section{Package Content}{\packageIndices{mx.api}}
\author{\packageAuthor{mx.api}}
\section{Maintainer}{\packageMaintainer{mx.api}}
