% tinyrox says don't edit this manually, but it can't stop you!
\name{mx_logout}
\alias{mx_logout}
\title{Log out of a Matrix session}
\usage{
mx_logout(session)
}
\arguments{
\item{session}{An "mx_session" object.}
}
\value{
Invisible NULL.
}
\description{
Invalidates the access token on the homeserver.
}
