% tinyrox says don't edit this manually, but it can't stop you!
\name{mx_room_leave}
\alias{mx_room_leave}
\title{Leave a room}
\usage{
mx_room_leave(session, room_id)
}
\arguments{
\item{session}{An "mx_session" object.}

\item{room_id}{Character. The room ID.}
}
\value{
Invisible NULL.
}
\description{
Leave a room
}
