% tinyrox says don't edit this manually, but it can't stop you!
\name{mx_room_members}
\alias{mx_room_members}
\title{List the members of a room}
\usage{
mx_room_members(session, room_id)
}
\arguments{
\item{session}{An "mx_session" object.}

\item{room_id}{Character. The room ID.}
}
\value{
Character vector of Matrix user IDs.
}
\description{
List the members of a room
}
