% tinyrox says don't edit this manually, but it can't stop you!
\name{mx_rooms}
\alias{mx_rooms}
\title{List rooms the user has joined}
\usage{
mx_rooms(session)
}
\arguments{
\item{session}{An "mx_session" object.}
}
\value{
Character vector of room IDs.
}
\description{
List rooms the user has joined
}
