% tinyrox says don't edit this manually, but it can't stop you!
\name{mx_whoami}
\alias{mx_whoami}
\title{Return the identity of the current session}
\usage{
mx_whoami(session)
}
\arguments{
\item{session}{An "mx_session" object.}
}
\value{
A list with user_id and device_id.
}
\description{
Return the identity of the current session
}
\examples{
\dontrun{
mx_whoami(s)
}
}
