% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{mx.api-package}
\alias{mx.api}
\alias{mx.api-package}
\title{mx.api: Minimal Matrix Client-Server API}
\description{
Base-R bindings for the Matrix Client-Server HTTP API, suitable for
talking to a Synapse homeserver. Two dependencies: curl and jsonlite.
End-to-end encryption is out of scope; use unencrypted rooms or a
separate crypto package.
}
\author{
\strong{Maintainer}: Troy Hernandez \email{troy@cornball.ai}
}
